requirements.txt
streamlit>=1.28.0
numpy>=1.23.0
opencv-python>=4.8.0
ultralytics>=8.0.0
streamlit-webrtc>=0.64.0
Pillow>=9.0.0
av>=10.0.0
torch>=2.0.0
torchvision>=0.15.0

